refactor : 카드 모양 enum Domain용과 View용으로 분리
