+++ src/main/java/blackjack/domain/Card.java
@@ -1,7 +1,5 @@
 package blackjack.domain;
 
-import blackjack.util.CardSuit;
-
 public class Card {
 
     private final CardNumber cardNumber;
