+++ src/main/java/blackjack/domain/CardSuit.java
@@ -1,0 +1,11 @@
+package blackjack.domain;
+
+import java.util.stream.Stream;
+
+public enum CardSuit {
+    SPADE,
+    DIAMOND,
+    HEART,
+    CLUB,
+    ;
+}
