+++ src/main/java/blackjack/domain/Cards.java
@@ -1,7 +1,6 @@
 package blackjack.domain;
 
 import blackjack.util.CardPickerGenerator;
-import blackjack.util.CardSuit;
 import java.util.ArrayList;
 import java.util.List;
