+++ src/main/java/blackjack/view/OutputView.java
@@ -3,8 +3,7 @@ package blackjack.view;
 import blackjack.domain.Participant;
 import blackjack.domain.Participants;
 import blackjack.domain.Player;
-import blackjack.domain.CardNumber;
-import blackjack.util.CardSuit;
+import blackjack.domain.CardSuit;
 import blackjack.util.WinningResult;
 import java.util.ArrayList;
 import java.util.List;
@@ -37,7 +36,7 @@ public class OutputView {
         List<String> cards = new ArrayList<>();
         for (int index = 0, end = participant.getReceivedCards().size(); index < end; index++) {
             ViewCardNumber cardNumber = ViewCardNumber.findCardNumber(participant.getCardNumber(index));
-            CardSuit cardSuit = CardSuit.getCardSuit(participant.getCardSuit(index));
+            ViewCardSuit cardSuit = ViewCardSuit.findCardSuit(participant.getCardSuit(index));
             cards.add(cardNumber.getName() + cardSuit.getCardSuitName());
         }
         System.out.print(String.join(", ", cards));
@@ -49,7 +48,7 @@ public class OutputView {
         List<String> cards = new ArrayList<>();
         for (int index = 0, end = participant.getReceivedCards().size(); index < end; index++) {
             ViewCardNumber cardNumber = ViewCardNumber.findCardNumber(participant.getCardNumber(index));
-            CardSuit cardSuit = CardSuit.getCardSuit(participant.getCardSuit(index));
+            ViewCardSuit cardSuit = ViewCardSuit.findCardSuit(participant.getCardSuit(index));
             cards.add(cardNumber.getName() + cardSuit.getCardSuitName());
         }
         System.out.print(String.join(", ", cards));
@@ -77,7 +76,7 @@ public class OutputView {
     private void printFirstDealerCards(final Participant participant) {
         if (participant.isDealer()) {
             ViewCardNumber cardNumber = ViewCardNumber.findCardNumber(participant.getCardNumber(FIRST_CARD));
-            CardSuit cardSuit = CardSuit.getCardSuit(participant.getCardSuit(FIRST_CARD));
+            ViewCardSuit cardSuit = ViewCardSuit.findCardSuit(participant.getCardSuit(FIRST_CARD));
             System.out.println(participant.getName() + " : " + cardNumber.getName() + cardSuit.getCardSuitName());
         }
     }
@@ -87,7 +86,7 @@ public class OutputView {
         List<String> card = new ArrayList<>();
         for (int i = 0; i < FIRST_HIT_COUNT; i++) {
             ViewCardNumber cardNumber = ViewCardNumber.findCardNumber(participant.getCardNumber(i));
-            CardSuit cardSuit = CardSuit.getCardSuit(participant.getCardSuit(i));
+            ViewCardSuit cardSuit = ViewCardSuit.findCardSuit(participant.getCardSuit(i));
             card.add(cardNumber.getName() + cardSuit.getCardSuitName());
         }
         System.out.printf(String.join(", ",card));
