+++ src/test/java/blackjack/domain/ResultReaderTest.java
@@ -2,7 +2,6 @@ package blackjack.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import blackjack.util.CardSuit;
 import blackjack.util.WinningResult;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
